% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{backtest_coint}
\alias{backtest_coint}
\title{Backtest_coint}
\usage{
backtest_coint(
  data,
  window_in = 250,
  lag_max = 5,
  z_in = 2,
  z_out = 0.5,
  lookback_z = 60,
  const_bps = 5,
  notional = 1000
)
}
\arguments{
\item{data}{data}

\item{window_in}{window_in}

\item{lag_max}{lag_max}

\item{z_in}{z_in}

\item{z_out}{z_out}

\item{lookback_z}{lookback}

\item{const_bps}{cost}

\item{notional}{notioanl}
}
\value{
tibble
}
\description{
backtester
}
