% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{backtest_coint_v2}
\alias{backtest_coint_v2}
\title{backtest coint v2}
\usage{
backtest_coint_v2(
  data,
  window_in = 250,
  lag_max = 5,
  z_in = 2,
  z_out = 0.5,
  lookback_z = 60,
  const_bps = 0,
  notional = 100
)
}
\arguments{
\item{data}{data}

\item{window_in}{a}

\item{lag_max}{b}

\item{z_in}{c}

\item{z_out}{d}

\item{lookback_z}{e}

\item{const_bps}{f}

\item{notional}{g}
}
\value{
data
}
\description{
bare rettelser
}
