% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_box.R
\name{best_box_spread}
\alias{best_box_spread}
\title{best_box_spread}
\usage{
best_box_spread(
  op_chain,
  contract_multiplier = 100,
  min_qty = 1,
  today = Sys.Date(),
  slippage_pad = 0
)
}
\arguments{
\item{op_chain}{chain}

\item{contract_multiplier}{100 som standard}

\item{min_qty}{1 std}

\item{today}{dato}

\item{slippage_pad}{slippage}
}
\value{
tibble
}
\description{
best_box_spread
}
