% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch data}
\usage{
fetch_data(conn)
}
\description{
This function fetches data from interactive brokers. Be aware that this only works while running the
IB-Gateway, see IBrokers documentation or the official IBKR documentation
}
